added conditional check on migration for obsolete table

+++ db/migrate/20140728155408_create_dialogues.rb
@@ -1,6 +1,8 @@
 class CreateDialogues < ActiveRecord::Migration
   def change
-    drop_table :dialogues
+    if ActiveRecord::Base.connection.table_exists? :dialogues
+      drop_table :dialogues
+    end
     
     create_table :dialogues do |t|
       t.string :title
